feat(dsh-provider-usage): 设置页多 tab 化——通知中心 #508 分段器同语言（#532）
整页一张 dou-set-card 大卡 + 顶部五分段器 tab（趋势/报告/用量/适配器/悬浮窗），
五区块一对一迁移行为不变。tab 状态纯 useState 内存态（不做 URL/存储持久化，
避免宿主路由事实源冲突）；窗格 keep-mounted + hidden 显隐（与旧全挂载请求时序
一致零回归，表单编辑态切 tab 保留；趋势图为 viewBox 自适应 SVG 无重测问题）。
settings.ts（806 行）拆 client/settings/ 目录：index（tab 骨架+数据面）/
usage/ui/providers/shared 五文件；locales 增五 tab 文案（zh/en）；style.css 增
dou-set-* 分段器样式（全走主题变量+浅色回退，移动端横向滚动兜底，focus-visible
环）。smoke 补契约断言：五 tab 键齐全、keep-mounted 语义、无持久化、结构类名
进产物。
