+++ packages/dsh-provider-usage/src/client/style.css
@@ -688,3 +688,56 @@
 .dou-provHead:hover { background: var(--dsw-alias-interactive-bg-hover, #f2f3f6); }
 .dou-float:hover { background: var(--dsw-alias-interactive-bg-hover, #f2f3f6); }
 }
+
+/* ================= #532 设置页多 tab（通知中心 #508 分段器同语言） =================
+ * 整页一张大卡 + 顶部五分段器（普通 button，role=navigation 不用 tablist——#402 决策）。
+ * 窗格 keep-mounted：hidden 属性切换显隐，组件实例不销毁（表单编辑态保留）。
+ * 颜色一律 --dsw-alias-* 主题变量 + 浅色回退（暗色由宿主变量保证）。 */
+.dou-set-card {
+  border: 1px solid var(--dsw-alias-border-l1, #e2e5ea);
+  border-radius: 12px;
+  overflow: hidden;
+  background: var(--dsw-alias-bg-base, #ffffff);
+  box-shadow: 0 1px 3px rgba(15, 23, 42, 0.05);
+}
+.dou-set-tabs {
+  display: flex;
+  gap: 4px;
+  padding: 10px 14px;
+  border-bottom: 1px solid var(--dsw-alias-border-l1, #e2e5ea);
+  background: var(--dsw-alias-bg-layer-1, #f5f6f8);
+  overflow-x: auto;
+  scrollbar-width: none;
+}
+.dou-set-tabs::-webkit-scrollbar { display: none; }
+.dou-set-tab {
+  display: inline-flex;
+  align-items: center;
+  gap: 6px;
+  min-height: 32px;
+  padding: 5px 14px;
+  white-space: nowrap;
+  border: 1px solid transparent;
+  border-radius: 8px;
+  background: transparent;
+  color: var(--dsw-alias-label-secondary, #5f6672);
+  font-size: 12px;
+  cursor: pointer;
+  font-family: inherit;
+}
+.dou-set-tabActive {
+  background: var(--dsw-alias-bg-base, #ffffff);
+  border-color: var(--dsw-alias-border-l1, #e2e5ea);
+  color: var(--dsw-alias-label-primary, #1f2329);
+  font-weight: 600;
+  box-shadow: 0 1px 3px rgba(15, 23, 42, 0.08);
+}
+.dou-set-tab:focus-visible {
+  outline: 2px solid var(--dsw-alias-state-info-primary, #4a7dd6);
+  outline-offset: 2px;
+}
+.dou-set-body { padding: 14px; }
+.dou-set-pane[hidden] { display: none !important; }
+@media (hover: hover) {
+  .dou-set-tab:hover { color: var(--dsw-alias-label-primary, #1f2329); }
+}
